refactor: Move `info-card` component from `empty-message` to its own folder
