+++ interfaces/portal/src/app/components/info-card/info-card.component.ts
